Very basic register allocation
incredibly naive implementation, somewhat bugtested for understanding of the problem. Need to reimplement in a more modular way since this method can't handle register spilling. Additionally, the code to keep track of which variables are in which registers (and alive or dead) can definitely be modularized to greatly increase simplicity and code readability.