rework of ASM for proper argument ordering, first ever code test!
The good news: the generated code ran and produced the correct results!
The bad news: the generated code duplicated the variable 'd' during the line 'd = d + 1'. I need to impelement a check to see whether a TAC line's destination is also an operand, and in that case figure out what to do. The simple lifetime checks for getting modifiable versions of variables aren't rigorous enough on their own, and this case is one that results in the variable being duplicated. The end result is that at the end of execution, one register contains an outdated value of the variable, and another register contains a 'dirty' updated value of the variable, and both registers are treated as 'live'. This will break even harder if more complex features are added because of the greedy search used to return the first register containing a variable with a matching name, so fixing it is priority #1 at the moment.