memory management, housekeeping
Most of the memory management issues are fixed. The main remaining thing is that the AST is horrendously inefficient. Nodes with the same value will receive individual malloc() and strcpy()s resulting in a duplication of data for identical names. Next up is to implement a dictionary so that parsed tokens will never be duplicated on the heap, and have all astNodes with the same value point to the dictionary entry.