Implemented function calls
The compiler can generate function calls for functions with arbitrary arguments, using the same expression parsing as it does for standard assignments. Still to do is push/pop operations to preserve registers modified within functions - this will likely require a rework of the assembly output to buffer it. This is due to the fact that register allocation happens as the code generator traverses the TAC, so the total number (and indices) of registers used isn't known until the code has all been generated.Register allocation during codegen is still pretty janky, and can't make use of the indirect addressing modes, instead opting to load everything into a register and do only register-register arithmetic. Additionally, I still haven't gotten around to register spilling, that will come after I get to work on an if statement and perhaps some other flow control depending on how hard it ends up being.