register state snapshot machinery
When entering/exiting if statements, the TAC now denotes when register states need to be saved (at a known point), when they should be restored from the most recent save, and when the most recent save should be popped. The savestates live on a stack which duplicates the 'struct registerState' info. TODO: implement the code to generate assembly to restore a given registerState** to an arbitrary registerState** by emitting the correct instructions to swap operands around and/or find them on the stack