register state restoring, unidentified bug with var duplication
in the current example source code there is a duplication of the variable 'result' into multiple registers. It doesn't seem like this is a result of any of the statestack stuff that was just implemented, it might be something to do with arithmetic or assignment generation. Will need to be looked into before continuing