Fix register state restoration by properly modifying stack pointer
previously, the add/sub on the stack pointer to make/free room for local variables happened *after* the callee would push registers and *before* the callee would pop registers. This would result in attempts to write local variables to the stack actually modifying the callee-saved register values. Switching the order in which the sp modifications are added to the ASM block fixed this.