implement new method for tracking variables and arguments on stack
instead of keeping track of index count of variables, each variable simply gets a type and a stack offset. Then, the total positive (for arguments) and negative (for local variables) stack offsets of each function are tracked and used for code generation. This will allow for differently sized variables than just a 16-bit 'var' at some point in the future