fix name mangling bug
when renaming variables, fix error by which their TAC representations wouldn't be renamed correctly. Now this operation is only undertaken when depth > 1 (in a subscope of a function) and is accomplished incrementally by iteratively converting all TAC lines.

This could be accomplished drastically more easily by simply swapping the dictionary pointer for that name to a different name value, but this would require a more drastic rework of the dictionary and break other references to the name, so is probably not worth it.